Move Fief admin helper class to own source file
